Silent warning

+++ techpack/video/msm/vidc/hfi_iris2.c
@@ -2,6 +2,7 @@
 /*
  * Copyright (c) 2019-2020, The Linux Foundation. All rights reserved.
  */
+#include <linux/kernel.h>
 
 #include "msm_vidc_debug.h"
 #include "hfi_common.h"
@@ -169,7 +170,7 @@ void __setup_ucregion_memory_map_iris2(struct venus_hfi_device *device, u32 sid)
 	__write_register(device, CPU_CS_VCICMDARG0_IRIS2,
 		(u32)device->iface_q_table.align_virtual_addr, sid);
 	__write_register(device, CPU_CS_VCICMDARG1_IRIS2,
-		(u32)((u64)device->iface_q_table.align_virtual_addr >> 32),
+		(u32)((uintptr_t)device->iface_q_table.align_virtual_addr >> 32),
 		sid);
 }
